some API changes
