+++ crates/crypto/examples/single_file_with_metadata.rs
@@ -1,10 +1,11 @@
 #![cfg(feature = "serde")]
 
 use sd_crypto::{
-	crypto::{Algorithm, StreamEncryptor},
+	crypto::StreamEncryptor,
 	header::{file::FileHeader, keyslot::Keyslot, metadata::MetadataVersion},
-	keys::hashing::{HashingAlgorithm, Params},
-	primitives::{Key, Salt, LATEST_FILE_HEADER, LATEST_KEYSLOT},
+	primitives::{
+		Algorithm, HashingAlgorithm, Key, Params, Salt, LATEST_FILE_HEADER, LATEST_KEYSLOT,
+	},
 	Protected,
 };
 use tokio::fs::File;
