+++ crates/crypto/src/header/metadata.rs
@@ -37,7 +37,10 @@ use crate::{
 
 use tokio::io::AsyncReadExt;
 
-use crate::{crypto::Algorithm, primitives::Nonce, Error, Result};
+use crate::{
+	primitives::{Algorithm, Nonce},
+	Error, Result,
+};
 
 use super::file::FileHeader;
